1. Fundamental functionality of RSS backend completed.

+++ app/src/main/java/com/maoyuchaxue/catfishclubnewsapp/data/db/CacheDBOpenHelper.java
@@ -28,6 +28,7 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
     public static final String FIELD_CATEGORY = "category";
     public static final String FIELD_TITLE = "title";
     public static final String FIELD_KEYWORDS = "keywords";
+    public static final String FIELD_TYPE = "type";
 
     private static final String NEWS_TABLE_CREATE = "create table if not exists " +
             NEWS_TABLE_NAME + " (" + FIELD_ID + " text primary key, " +
@@ -44,7 +45,8 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
             FIELD_CRAWL_SRC + " text, " +
             FIELD_CATEGORY + " text, " +
             FIELD_TITLE + " text, " +
-            FIELD_KEYWORDS + " text);";
+            FIELD_KEYWORDS + " text, " +
+            FIELD_TYPE + " integer not null default 0);";
 
 
     public static final String RESOURCES_TABLE_NAME = "resources_cache";
@@ -69,7 +71,8 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
             FIELD_LANG + " text, " +
             FIELD_SRC + " text, " +
             FIELD_PICTURES + " text, " +
-            FIELD_VIDEO + " text);";
+            FIELD_VIDEO + " text," +
+            FIELD_TYPE + " integer not null default 0);";
 
     public static final String RSS_TABLE_NAME = "rss_cache";
     public static final String FIELD_DESC = "description";
@@ -81,6 +84,21 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
             FIELD_TITLE + " text, " +
             FIELD_LINK + " text, " +
             FIELD_DESC + " text);";
+
+    public static final String RSS_NEWS_TABLE_NAME = "rss_news_cache";
+//    private static final String RSS_NEWS_TABLE_CREATE =
+    private static final String RSS_NEWS_TABLE_CREATE = "create table if not exists " +
+            RSS_NEWS_TABLE_NAME + " (" +
+            FIELD_ID + " text primary key, " +
+            FIELD_INTRO + " text, " +
+            FIELD_CATEGORY_TAG + " integer, " +
+            FIELD_AUTHOR + " text, " +
+            FIELD_TITLE + " text, " +
+            FIELD_URL + " text, " +
+            FIELD_LANG + " text, " +
+            FIELD_SRC + " text, " +
+            FIELD_PICTURES + " text, " +
+            FIELD_VIDEO + " text);";
 
 
     private CacheDBOpenHelper(Context context){
@@ -111,11 +129,16 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
         if(oldVersion < 4)
             db.execSQL(BOOKMARK_TABLE_CREATE);
 
-        if(oldVersion < 3 || newVersion >= 7){
+        if(oldVersion < 3 || (newVersion >= 7 && oldVersion < 7)){
             db.execSQL("drop table if exists " + NEWS_TABLE_NAME + ";");
             db.execSQL(NEWS_TABLE_CREATE);
 //            db.execSQL("alter table " + NEWS_TABLE_NAME +
 //            "add column " + FIELD_TITLE + " text;");
+        }
+
+        if(newVersion == 8){
+            db.execSQL("drop table if exists " + BOOKMARK_TABLE_NAME + ";");
+            db.execSQL(BOOKMARK_TABLE_CREATE);
         }
 
         if(oldVersion < 5){
@@ -132,6 +155,14 @@ public class CacheDBOpenHelper extends SQLiteOpenHelper {
 
         if(oldVersion < 8)
             db.execSQL(RSS_TABLE_CREATE);
+
+        if(newVersion == 8){
+            db.execSQL("alter table " + NEWS_TABLE_NAME + " add column " +
+                FIELD_TYPE + " integer not null default 0;");
+            db.execSQL("alter table " + BOOKMARK_TABLE_NAME + " add column " +
+                    FIELD_TYPE + " integer not null default 0;");
+        }
+
     }
 
     @Override
